Kit_of_ShareableLinks + living NPC list links (QST-0021.6, QST-0011)
- The ~200-line inline share/hash script moves to
  app/static/js/shareable-links.js, served by
  AtlasVenustas/Kit_of_ShareableLinks.shareable_links_head_tags().
  Boundary per ticket: encode/decode stays in app/character_url.py;
  CharacterPathRedirectASGI stays in the entrypoint (ASGI, not
  presentation).
- NPC list entries stop being href='#': each row is an action link
  that rebuilds that exact NPC - same race, archetype, level, and
  seed, full detail instead of light - and opens the NPC page.
  NPC_LIST_SIZE constant shared by generator, renderer, observers.

shiny_app.py head assembly is now the thin form the parent ticket
sketched: style_tag(), *loader_head_tags(), *tablet_head_tags(),
*masonry_head_tags(), *shareable_links_head_tags().
